Initial implementation using Hibernate ORM and Authentication.
The first working code using hibernate to persist the Userd object
to the database. The User entity is called Userd because User is a
keyword in Postgresql and to avoid unnecesary configuration it is easier to just
use Userd for User(database) at the Model.

For reference this is how to use the jpa: http://www.objectdb.com/java/jpa/query

Also, the first working code that implements authentication. Using the
@Security.Authenticated(Secured.class) we can make sure that some
resources are only displayed to logged in users.

I followed this guide: http://www.playframework.com/documentation/2.1.0/JavaGuide4 along with this https://github.com/yesnault/PlayStartApp to implement the authentication.