Implemented authentication with jdbc.
Fixed a compile time error in UserController
Changed the session key to be "id" rather than email.
Removed JPA imports in controllers
Implemened some of the main  functionallity of the Userd class.
Set up static class variables that represent the column names in the
tables.

This is a pretty big commit